Implement AccountScreen class with chatbot functionality and update Account model to ensure balance is a double

+++ curso_dart_dominando_assincronismo_e_criando_comunicacao_com_apis/dart_assincronismo/lib/models/account.dart
@@ -16,7 +16,7 @@ class Account {
       id: map['id'],
       name: map['name'],
       lastName: map['lastName'],
-      balance: map['balance'],
+      balance: map['balance'].toDouble(),
     );
   }
